# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Prova Azienda
  version: 1.0.0
paths:
...
 /project: 
   post:
      summary: Add a new Project 
      description: Add a new Project
      operationId: addProject
      requestBody:
        description: Create a new Project
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
          application/xml:
            schema:
              $ref: '#/components/schemas/Project'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Project'
        '405':
          description: Invalid input 
 /project/{projectId}:   
  get:
      summary: Find Project by ID
      description: Returns a single Project
      operationId: getProjectById
      parameters:
        - name: projectId
          in: path
          description: ID of project to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
  post:
      summary: Updates a project  with form data
      description: ''
      operationId: updateProjectWithForm
      parameters:
        - name: projectId
          in: path
          description: ID of project that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: projectName
          in: query
          description: Name of project that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input        
  delete:
      summary: Deletes a project
      description: delete a project
      operationId: deleteProject
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: projectId
          in: path
          description: Project id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid project value 
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...