# target/generated-sources/openapi/src/gen/java/main/org/openapitools/api/ProjectApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Project;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-09-15T18:52:56.718999500+02:00[Europe/Rome]")
@Validated
@Tag(name = "project", description = "the project API")
public interface ProjectApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /project : Add a new Project
     * Add a new Project
     *
     * @param project Create a new Project (required)
     * @return Successful operation (status code 200)
     *         or Invalid input (status code 405)
     */
    @Operation(
        operationId = "addProject",
        summary = "Add a new Project",
        description = "Add a new Project",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Project.class))
            }),
            @ApiResponse(responseCode = "405", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/project",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<Project> addProject(
        @Parameter(name = "Project", description = "Create a new Project", required = true) @Valid @RequestBody Project project
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"name\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    String exampleString = "<null> <id>123456789</id> <name>aeiou</name> </null>";
                    ApiUtil.setExampleResponse(request, "application/xml", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /project/{projectId} : Deletes a project
     * delete a project
     *
     * @param projectId Project id to delete (required)
     * @param apiKey  (optional)
     * @return Invalid project value (status code 400)
     */
    @Operation(
        operationId = "deleteProject",
        summary = "Deletes a project",
        description = "delete a project",
        responses = {
            @ApiResponse(responseCode = "400", description = "Invalid project value")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/project/{projectId}"
    )
    default ResponseEntity<Void> deleteProject(
        @Parameter(name = "projectId", description = "Project id to delete", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId,
        @Parameter(name = "api_key", description = "", in = ParameterIn.HEADER) @RequestHeader(value = "api_key", required = false) String apiKey
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId} : Find Project by ID
     * Returns a single Project
     *
     * @param projectId ID of project to return (required)
     * @return successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Employee not found (status code 404)
     */
    @Operation(
        operationId = "getProjectById",
        summary = "Find Project by ID",
        description = "Returns a single Project",
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Project.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Employee not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}",
        produces = { "application/json", "application/xml" }
    )
    default ResponseEntity<Project> getProjectById(
        @Parameter(name = "projectId", description = "ID of project to return", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"name\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    String exampleString = "<Project> <id>123456789</id> <name>aeiou</name> </Project>";
                    ApiUtil.setExampleResponse(request, "application/xml", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /project/{projectId} : Updates a project  with form data
     * 
     *
     * @param projectId ID of project that needs to be updated (required)
     * @param projectName Name of project that needs to be updated (optional)
     * @return Invalid input (status code 405)
     */
    @Operation(
        operationId = "updateProjectWithForm",
        summary = "Updates a project  with form data",
        description = "",
        responses = {
            @ApiResponse(responseCode = "405", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/project/{projectId}"
    )
    default ResponseEntity<Void> updateProjectWithForm(
        @Parameter(name = "projectId", description = "ID of project that needs to be updated", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId,
        @Parameter(name = "projectName", description = "Name of project that needs to be updated", in = ParameterIn.QUERY) @Valid @RequestParam(value = "projectName", required = false) String projectName
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}